refactor: split runs evidence command
